get_integer_operand의 반환 타입을 uint32_t로 수정

+++ include/Instruction.hpp
@@ -38,7 +38,7 @@ public:
 
   Opcode get_opcode();
   const std::string &get_string_operand();
-  uint64_t get_integer_operand();
+  uint32_t get_integer_operand();
 
 private:
   uint32_t line_number_;
